Reworked deadline system to be way more streamlined and made it more accurate in the process + bugfix
Separated a lot of the code into a separate file 'date.js'. The old code in 'content.js' was super messy and hard to read and didn't even work that well (sometimes it was off by a minute or two).

Bugfix:
Extension couldn't observe for change in list of assignments if there were no assignments (this would happen when you had no upcoming assignments)