code cleanup
removed old debugging statements
removed old print statements
removed old comments
added additional comments for context
fixed a bug in getTranslatableDateString() where it would always remove the last 4 characters of the date string to remove the final "and". Problem is that the equivalent of "and" in other languages can be of a different length and therefore the amount to subtract is different.