removed old language system + fixed race condition
the build() function inside of the fetch().then() function was asynchronous (and therefore returned a promise), but the code never awaited that promise, occasionally causing a race condition.
I had to completely tear the read function (that called fetch().then()) apart since I couldn't await for the build() function inside of the fetch().then() callback.